refactor cropper comp, update NuGet packages
