+++ examples/Cropper.Blazor.MAUI.Net8/Cropper.Blazor.MAUI.Net8.csproj
@@ -62,7 +62,7 @@
 		<PackageReference Include="Microsoft.Maui.Controls.Compatibility" Version="8.0.100" />
 		<PackageReference Include="Microsoft.AspNetCore.Components.WebView.Maui" Version="8.0.100" />
 		<PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="8.0.1" />
-		<PackageReference Include="MudBlazor" Version="8.7.0" />
+		<PackageReference Include="MudBlazor" Version="8.13.0" />
 	</ItemGroup>
 
 	<ItemGroup>
