+++ examples/Cropper.Blazor.Server.Net8/Cropper.Blazor.Server.Net8.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="MudBlazor" Version="8.7.0" />
+    <PackageReference Include="MudBlazor" Version="8.13.0" />
   </ItemGroup>
 
 	<ItemGroup>
