+++ src/Cropper.Blazor/Cropper.Blazor.Shared/Cropper.Blazor.Shared.csproj
@@ -6,7 +6,7 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="MudBlazor" Version="8.11.0" />
+		<PackageReference Include="MudBlazor" Version="8.13.0" />
 	</ItemGroup>
 
 </Project>
