+++ src/Cropper.Blazor/Cropper.Blazor/.config/dotnet-tools.json
@@ -3,7 +3,7 @@
   "isRoot": true,
   "tools": {
     "excubo.webcompiler": {
-      "version": "4.0.11",
+      "version": "4.0.33",
       "commands": [
         "webcompiler"
       ]
